Abandon 'TypeScript'
Based on the work by Marston (6to5), the code base now switches to ES6.
TypeScript sources are kept for fear that someone may need to reference
them.

No one should build on those '.ts' files and the corresponding part in
'gulpfile' are commented out and soon be removed entirely.

Reminder: running `gulp` is sufficient to start Express with nodemon.

Change-Id: Ie1d1469d3b23368174383a289dfd3088bfdd687d