Define an email job for scheduling
The email job is used for saving the email details (from MongoDB) to the
queue backed by Redis. The processing logic `processFn` is external to
the job itself (passed thru constructor) to avoid the job from coupling
with any collaborating object(s).

Interface may be defined for the job if a pattern emerges.

Please run `npm i` after this update.

Others:
- Schemas now should all use 'createdAt' (instead of 'createAt')
- Fix 'email' collection's 'template-*' logic
- Break down 'watch' gulp task in case you can watch only 'js' assets

Jira: WEB-152

Change-Id: I0f3306122ad6a552a59c8c5f15a73823bb4053eb