Project architecture change
The stack has now become 'Fluxible' (isomorphic stuffs) +
'react-router' (component routing). The whole idea is shamelessly copied
from the reference project mentioned below.

Data fetching could be defined in the `fetchData` method with the
component, and since RR supports nested routing, any children should be
able to 'share' the data fetched by it parent.

Also, we are using a forked branch of RR (before v1.0 is coming). See
the reason in the reference below
But the migiration path should not be a big deal.

Reference:

- [Reference Project](https://github.com/nicolashery/example-isomorphic-one)
- [Why use forked branch](https://github.com/rackt/react-router/pull/590)
- [Forked Details](https://github.com/bobpace/react-router/commit/a805f919c6deaa50e2a1f4a5a69a18f8de77f8b1)
- [React Router](https://rackt.github.io/react-router/)

Others:

- remove 'fluxible-plugin-*' dependencies but keep 'fluxible-router' for
  now to ease migration

Minor:

- Add new components for proofing this transition
- include 'immutable' for easier migration
- remove 'hiredis' to allow using latest version of iojs
- remove remaining obsolete path for AngularJS
- [git] ignore generated diagrams

TODO:

- 'Wrapper' component should be splitted into 2 and replace the
  'App' component for now we can handle authentication state with the
  help of 'willTransitionTo'

Change-Id: Ia7cb4eb901ce35a62bc2ba7e3a06271979ebda83