Consolidate URL reference for the application itself
Add an URL utility functions to ease the URL resolution. The urls used
in the "Api" object need to be factored out. The `this._getHost` is the
preferred approach for now due to its deferred evaluation of the env
variable(s) resolution.

The following environment variables are added:

- APP_HOST
- APP_PORT

Change-Id: I7755a5c0839e5f238b9e852beb4d721864ba3265
Jira: WLP-73