Fixed invalid monthly active user stats
The correct way is to query the last day of month with a
timeWindow of `month`. This will give a result of number of
unique active user within a month up to the last day of month.

On development server, the current month (e.g. yesterday) data
may not be available as the data computing script is only run
when required.

Change-Id: I51633ecdf0502afcb04aec3d20482b6d50a07e0c
Jira: WLP-512