Added Call Stats API
The structure follows other stat request and they will be
refactored all together after wrapping up as the structure now
is very difficult to test

BREAKING CHANGES:
The dataProvider request handlers now support multiple API endpoints
where the format is {domain1}:{port1},{domain2}:{port2}. The first
index will be default and it should be the most powerful one.

the sendRequest method takes an optional param of loadBalanceIndex
and it will always use the first endpoint if loadBalanceIndex is 0.

to simulate the load balancing, you may separate the query by using
splitQuery method in requestHelper, map the output queries array and
pass the query and index to the sendRequest method.

Change-Id: If31ccb1773a4b0a53c4dccb7a4bb1ad5b3b3c336
Jira: WLP-450