Replaced ASR Chart with CombinationChart Component
The CombinationChart component is duplicated from LineChart component,
but it allows user to define the type of series rather than
pre-setting it as `line`. It also enable user to configure the
visibility of the Chart Legend with props.

When doing code review, you may not have to go through the whole
CombinationChart component, as it shares a large amount of logic
with that of LineChart. You may pay specific attention to
`addSeries` and `drawChart` method.

I rejected to use legend alignment with layout (which is vertical)
which would induce a hacky x,y positioning.

Change-Id: I67eb63e53421c370a9d1978fbabeca3372721bf4
Jira: WLP-578, WLP-581
See: WLP-589