Fixed Distorted Chart in Call Overview
this bug is due to the uncleared(old) data in store,
while this situation is commonly seen all over the project.

clearing the data in store in componentWillUnmount will be the
most efficient way to fix the bugs, until we figure out a
proper way to preserve the data in store.

Change-Id: Idd20ad513287c34781a15a8cbe5e5367c3418976
Jira: WLP-596