Fixed Invalid Device Stats in EndUserOverview
In commit 5c8c0e, the time frame picker changed from getting
`n to n-6` to `n-1 to n-7` so the device total is out-sync with
user total. Now the device total no longer related to the selected
time frame; however, the layout is not updated yet so the device
total is still placed within the time frame section.

Change-Id: I5d64588fadd422a0cd6049194eb475c862aa0653