Fixed inaccurate Line Charts in Verification Overview
The state management in Verification Overview is messy. It very
much needs to be refactored.

The previous error is caused by inconsistent data with x-Axis. Change
in selected time frame will make the xAxis to change immediately, but
not the data(series), causing an inaccurate presentation in charts.

Change-Id: I6ddc60139d7537525af78b44068b06aa2c707c96
Jira: WLP-650, WLP-658