Implemented cancellable requests for CALL OVERVIEW
Upon ACTION_START, the request will be pushed into STORE, and
upon CLEAR_ACTION, the requests in store will be aborted and cleared.

The current design assume that ACTION and request are 1-to-1 match.
When ACTION_START is called before the same previous action ended,
the related pending request will be aborted, and overwritten by the
new request.

Change-Id: I613ba55acb475e547a07fc053b742728d13d8a2d
Jira: WLP-661