Changed webpack to use 'source-map'
I am not sure if it should be used in production, as the build speed
is slow. However, it is used for m800-online-platform-enterprise and
mwidget-web-client.

With this setting, we can see the original source in browser
inspector for debugging.

Change-Id: If4f2fe4453b709bdb83cb8819aba7787c7ea31a8
See: http://webpack.github.io/docs/configuration.html#devtool
Jira: WLP-568