Wrapped the react-related dependencies upgrade
With this changes, you will be able to start the application, but the
browsable sections are just SIGN IN, CALLS, and END USERS pages.

It is known that the sign out function does not work anymore, as it
failed to clear the cookies on browser. The auth and session flow will
be refactored as discussed, so this issue is ignored in the upgrade
process.

You may regard the dependencies upgrade is finished here. All the other
changes will be done when the corresponding components are being
refactored.

Change-Id: I77f61483a515cd723d5c86874ebb5939f832bebd
Jira: WLP-566, WLP-684, WLP-169, WLP-687