Implemented timeout workaround on ApiClient
`timeout` option is not included in the documentation.

It works on server-side but not client side if `timeout` is
added to the `options` argument.

So a workaround is done as suggested in an issue in order to
provide a more consistent outcome before the `timeout` option
is officially implemented.

@see: http://github.github.io/fetch/

Change-Id: Idbd91d03acfd97981293fab7ce80ddd172f2a101
Jira: WLP-688