Made Panel component accepting ReactElement as caption
I found that duplicate key problem always happen to the string props
in Common Components. I regard this change as one of the solution.

another way is to define the common messages (application level) in
the outermost wrapper component, and ask everyone to refer to that
before defining messages in their own component.

Change-Id: I88692e05967d33502c05ef39d18847e27c5e9629
Jira: WLP-465