Fix Jenkins merge fail of installing not compatible version of dependency
The latest version of browser-sync requires browser-sync-ui 0.6.0, but the latest build of browser-sync-ui is failed. As a result, browser-sync-ui 0.6.0 was not published correctly. Requiring a non-exist version break the build during merge.

It is a quick fix by limiting the version number that references to the version of successful until their build is back to normal.

Change-Id: Icce5c6121702ed671c672773faa5639b8e501983