Apply the open id connect
Removed the signin components, updated on how to login using open id connect by updating the auth stores, changing the auth routes and passport strategy.
Added open id passport strategy and applied the oidc client to request and introspect the tokens.

Sign in will make use of passport strategy to redirect to the IAM service to login, sign out will redirect to the end session end point in the sign out action.
When accessing the application, it will check the authority in the app/routes.js and redirect to /sign-up which initialize the open id authorize flow.

JIRA: WLP-955

Change-Id: I080f927056d8486951de5305dc16f353c7aa8e8b