Remove the role concept.
Since it will be replaced by new access control. Role/access permission will be dynamic and check for every routes
rather than by role to determine the action.

Also the purpose for that in the current implementation is mainly for the acl middleware and link. It should not create critical problems after the changes.

Change-Id: I675a533ad3f99c5f0ebd496fa70f4db16c62e04c